Move utils into submodule
